third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCIceCandidate-constructor-expected.txt

This is a testharness.js-based test.
PASS new RTCIceCandidate()
FAIL new RTCIceCandidate({}) assert_throws: function "() => new RTCIceCandidate({})" threw object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty." ("TypeMismatchError") expected object "TypeError" ("TypeError")
FAIL new RTCIceCandidate({ ... }) with manually filled default values assert_throws: function "() => new RTCIceCandidate({
        candidate: '',
        sdpMid: null,
        sdpMLineIndex: null,
        ufrag: undefined
      })" threw object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty." ("TypeMismatchError") expected object "TypeError" ("TypeError")
FAIL new RTCIceCandidate({ sdpMid: null, sdpMLineIndex: null }) assert_throws: function "() => new RTCIceCandidate({
        sdpMid: null,
        sdpMLineIndex: null
      })" threw object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty." ("TypeMismatchError") expected object "TypeError" ("TypeError")
FAIL new RTCIceCandidate({ candidate: '' }) assert_throws: function "() => new RTCIceCandidate({
        candidate: ''
      })" threw object "TypeMismatchError: Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty." ("TypeMismatchError") expected object "TypeError" ("TypeError")
FAIL new RTCIceCandidate({ candidate: null }) assert_throws: function "() => new RTCIceCandidate({
        candidate: null
      })" did not throw
FAIL new RTCIceCandidate({ ... }) with valid candidate string only assert_throws: function "() => new RTCIceCandidate({
        candidate: candidateString
      })" did not throw
FAIL new RTCIceCandidate({ sdpMid: 'audio' }) Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ sdpMLineIndex: 0 }) Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ sdpMid: 'audio', sdpMLineIndex: 0 }) Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ candidate: '', sdpMid: 'audio' } Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ candidate: '', sdpMLineIndex: 0 } Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ ... }) with valid candidate string and sdpMid assert_equals: expected (object) null but got (number) 0
FAIL new RTCIceCandidate({ ... }) with invalid candidate string and sdpMid assert_equals: expected (object) null but got (number) 0
FAIL new RTCIceCandidate({ ... }) with non default value for all fields assert_equals: expected (string) "test" but got (undefined) undefined
FAIL new RTCIceCandidate({ ... }) with invalid sdpMid Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
FAIL new RTCIceCandidate({ ... }) with invalid sdpMLineIndex Failed to construct 'RTCIceCandidate': The 'candidate' property is not a string, or is empty.
Harness: the test ran to completion.
